+++ src/main/java/strategy/updater/observer/ClearLocalStackExitObserver.java
@@ -22,7 +22,7 @@ public class ClearLocalStackExitObserver implements DirectionUpdaterObserver {
     @Override
     public void notifyOfStateChange(SkelatalAgent va) {
         if (key != null && va.getStackedPosition(key) != null) {
-            Stack<float[]> stack = va.getStackedPosition(key);
+            Stack<short[]> stack = va.getStackedPosition(key);
             stack.clear();
         }
     }
